Separate twitter link search into new file
