+++ src/TextifyBot.py
@@ -16,6 +16,7 @@ from PIL import Image
 import time
 import math
 import TextifyTranslate
+import TextifyTwitter
 from googlesearch import search
 
 
@@ -38,7 +39,7 @@ TESSERACT_PATH = 'C:\\Program Files\\Tesseract-OCR\\tesseract.exe'
 CHECKER = True # enables posting to Reddit
 TRANSLATE_FLAG = '!Translation'
 CV_KEYWORD = "!describe" # keyword for providing a description of an image
-TWITTER_FLAG = '!Twitter';
+TWITTER_FLAG = '!Twitter' #keyword for twitter link search request
 
 
 
@@ -107,9 +108,6 @@ def processMention(mention):
                     mention.reply("Transcription was unable to identify any text within the image")
                 else:
                     makeReply(mention, result)
-        else:
-            if CHECKER:
-                mention.reply("No URL(s) found")
 
 # - Post's subreddit must not be in blacklist
 # - Post's subreddit must not be NSFW (+18)
@@ -134,13 +132,13 @@ def makeReply(mention, transcriptions):
     #check for twitter flag
     if mention.body.find(TWITTER_FLAG) >=0:
         print("Twitter flag found, searching google for link...")
-        twitterLink = googleSearch(response)
+        twitterLink = TextifyTwitter.googleSearch(response)
         if twitterLink != None:
             response = response + '\n\n' + twitterLink
         else:
             print("unable to find link")
             response = response + '\n\nUnable to find associated twitter link'
-                
+
     #check for translation flag
     if mention.body.find(TRANSLATE_FLAG) >=0:
         for langCode in TextifyTranslate.LANGUAGE_CODE:
@@ -279,53 +277,6 @@ def escapeMarkdown(str1):
                     str1 = newString
                     x = index1 + 2
     return newString
-
-
-# If twitter keyword is used, search google to see if link to tweet can be found
-def googleSearch(response):
-    for link in search(response, tld="com", num=10, stop=1, pause=2):
-        if link.find('twitter') >= 0:
-            print('found twitter link: ' + link)
-            return 'Twitter link found: ' + link
-        else:
-            return getTwitterUsernames(response)
-
-    #have to chek here as well in case the search returns no results
-    #if the search for the entire tweet returns no results, which it will occassionally with jumbled transcriptions,
-    #the above for loop will immediately exit
-    if getTwitterUsernames(response) == None:
-        return None
-    else:
-        return getTwitterUsernames(response)
-
-
-#if google search is unable to find specific tweet, will attempt to find the twitter user page
-def getTwitterUsernames(response):
-    twitterUsernames = extractUsername(response)
-    usernameList = ''
-    for username in twitterUsernames:
-        if username != None:
-            for link in search(username, tld="com", num=10, stop=1, pause=2):
-                if link.find('twitter') >= 0:
-                    usernameList = usernameList + '\n\n' + link
-                    print('found twitter user page: ' + link)
-
-    if usernameList == None:
-        return None
-    else:
-        return 'Twitter user(s) found: ' + usernameList
-
-
-# If tweet is not able to be found, extract username from tweet and return it to search for it
-def extractUsername(response):
-    responseSplit = response.split()
-    usernames = []
-
-    for word in responseSplit:
-        if word.find('@') >= 0:
-            usernames.append(word)
-
-    return usernames
 
 
 # Main driver code
